async improvements:
- Rename `Curl_doh_is_resolved()` to `Curl_doh_take_result()` to
  make clear this can only be called once successfully and deliver
  the result.
- No longer keep the resolved dns entry at the async struct. It is
  delivered to the caller of `Curl_doh_take_result()`  and not
  kept around.
- use `uint16_t` for `port`, `uint8_t` for `ip_version` everywhere
- remove `Curl_resolv_timeout()` and add a `timeout_ms` parameter
  to `Curl_resolv()`
- hostip.c: split the alarm timeout resolve code from the "normal"
  resolve code for clarity

+++ lib/vtls/vtls.c
@@ -1248,7 +1248,7 @@ CURLcode Curl_ssl_peer_init(struct ssl_peer *peer,
   {
     ehostname = cf->conn->host.name;
     edispname = cf->conn->host.dispname;
-    peer->port = cf->conn->remote_port;
+    peer->port = (uint16_t)cf->conn->remote_port;
   }
 
   /* hostname MUST exist and not be empty */
